requirements.txt
# requirements.txt
pyyaml
beautifulsoup4
lxml
pytest
pytest-mock
pyfakefs
